Refactored: Settings specs.
